Tidy up JNI LatLng refs

+++ android/java/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxgl/lib/constants/MathConstants.java
@@ -1,8 +1,8 @@
 package com.mapbox.mapboxgl.lib.constants;
 
 public interface MathConstants {
-    public static final float DEG2RAD = (float) (Math.PI / 180.0);
-    public static final float RAD2DEG = (float) (180.0 / Math.PI);
+    public static final double DEG2RAD = (Math.PI / 180.0);
+    public static final double RAD2DEG = (180.0 / Math.PI);
 
-    public static final float PI = (float) Math.PI;
+    public static final double PI = Math.PI;
 }
